Moves logic to logic.py module
