+++ src/security/templatetags/securitytags.py
@@ -3,7 +3,6 @@ from django import template
 from core.middleware import GlobalRequestMiddleware
 from security import logic
 from submission import models
-from utils import setting_handler
 
 register = template.Library()
 
@@ -93,35 +92,11 @@ def is_preprint_editor(context):
     return request.user.is_preprint_editor(request)
 
 
-def can_see_pii(request, article):
-    # Before doing anything, check the setting is enabled:
-    se_pii_filter_enabled = setting_handler.get_setting(
-        setting_group_name='permission',
-        setting_name='se_pii_filter',
-        journal=article.journal,
-    ).processed_value
-
-    if not se_pii_filter_enabled:
-        return False
-
-    # Check if the user is an SE and return an anonymised value.
-    # If the user is not a section editor we assume they have permission
-    # to view the actual value.
-    stages = [
-        models.STAGE_UNASSIGNED,
-        models.STAGE_UNDER_REVIEW,
-        models.STAGE_UNDER_REVISION,
-    ]
-    if request.user in article.section_editors() and article.stage in stages:
-        return True
-    return False
-
-
 @register.filter
 def se_can_see_pii(value, article):
     request = GlobalRequestMiddleware.get_current_request()
 
-    if can_see_pii(request, article):
+    if logic.can_see_pii(request, article):
         return 'Value Anonymised'
     else:
         return value
@@ -131,7 +106,7 @@ def se_can_see_pii(value, article):
 def can_see_pii_tag(context, article):
     request = context.get('request')
 
-    if can_see_pii(request, article):
+    if logic.can_see_pii(request, article):
         return False
     else:
         return True
